% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_highs.R
\name{city_highs}
\alias{city_highs}
\title{High Temperatures by Month}
\usage{
city_highs(city_name)
}
\arguments{
\item{city_name}{String giving the name of the city, Possible values
are: Auckland, Mumbai, Beijing, Chicago, and San Diego.}
}
\description{
High temps by month (violin plots) for several cities.
}
\examples{
city_highs("Auckland")
city_highs("Mumbai")
}
